+++ src/Html/Builder.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace ElfSundae\Laravel\Datatables\Html;
+namespace ElfSundae\Laravel\DataTables\Html;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
-use Yajra\Datatables\Html\Builder as BaseBuilder;
+use Yajra\DataTables\Html\Builder as BaseBuilder;
 
 class Builder extends BaseBuilder
 {