+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# Changelog
+# Release Notes
 
-All notable changes to `laravel-datatables` will be documented in this file.
+## 2.x.x ()
 
-## 1.0.0 - 2017/06/18
+- Update `yajra/laravel-datatables-oracle` to v8.x
+- Rename namespace From `ElfSundae\Laravel\Datatables` to `ElfSundae\Laravel\DataTables`
+
+## 1.0.0 (2017-06-18)
 
 - Initial release